% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome-hierachy.R
\name{add_tree_reactome}
\alias{add_tree_reactome}
\title{Add Reactome hierarchical structure to rows of a matrix or data.frame of data for Reactome pathways}
\usage{
add_tree_reactome(sl_table,
  store_temp = (file.exists("ReactomePathwaysRelation.txt") &
  file.exists("ReactomePathways.txt")))
}
\arguments{
\item{sl_table}{Input data: matrix or data.frame with a row for each Reactome pathway. The first column ("Probe") must be the names of Reactome pathways and the second column ("Reactome ID") is used if given.}

\item{store_temp}{logical. Whether or not to store the pathway structure data. Defaults to FALSE (sourcing files from URL) unless files found in working directory.}
}
\value{
The sum input data sorted by Reactome pathways with an additional column (left) for annotation.
}
\description{
Add Reactome hierarchical structure to rows of a matrix or data.frame of data for Reactome pathways
}

